requirements.txt
streamlit
google-generativeai
google-cloud-storage
numpy
scikit-learn
torch
sentence-transformers
scikit-learn
google-auth
google-auth-oauthlib
matplotlib